+++ packages/expo-ui/ios/MaskView.swift
@@ -22,7 +22,7 @@ internal struct MaskView: ExpoSwiftUI.View {
 
   @ViewBuilder
   private var baseContent: some View {
-    ForEach(props.children?.withoutSlots() ?? [], id: \.id) { child in
+    ForEach(props.children?.withoutSlots() ?? [], id: \.childIdentity) { child in
       let view: any View = child.childView
       AnyView(view)
     }