+++ packages/expo-ui/ios/OverlayView.swift
@@ -19,7 +19,7 @@ internal struct OverlayView: ExpoSwiftUI.View {
 
   @ViewBuilder
   private var baseContent: some View {
-    ForEach(props.children?.withoutSlots() ?? [], id: \.id) { child in
+    ForEach(props.children?.withoutSlots() ?? [], id: \.childIdentity) { child in
       let view: any View = child.childView
       AnyView(view)
     }