+++ packages/expo/ios/Expo/ExpoAppSceneDelegate.swift
@@ -2,6 +2,9 @@
 
 import Foundation
 import React
+#if canImport(ExpoObjC)
+import ExpoObjC
+#endif
 
 #if os(iOS) || os(tvOS)